EntropyFS: Phase 2 persistent store — crash-consistent immutable object store
Store core (segments, dual-superblock roots, COW B-tree indexes, inode
tree, directory tree, extent tree, transactions, GC, recovery), now
correct and fully green:

- Transaction commit protocol with crash-court injection points at every
  durability boundary; crash matrix at all 6 boundaries passes with
  pre/post-state admissibility verified after reopen.
- Fix root-dir inode creation at mkfs (mounts require inode 2 to exist).
- Fix read path: holes materialize as ZERO and reads clip to inode.size;
  hole-skip via next-extent lookup; output indexed by absolute position.
- Fix truncate: trailing partial extents are re-encoded and trimmed so no
  extent extends past the file size.
- Fix GC: dead index entries removed after victim deletion; record
  envelopes preserve flags/materialized_len when compacted.
- Fix COW B-tree: changed nodes get new content ids and propagate upward;
  order is validated at decode time (corrupt/mismatched nodes rejected).
- Fix cost accounting: Data-object payload bytes now counted in
  persisted_bytes (RAW/RANS were scored ~free, so rANS could never win).
- Fix segment reopen: torn tails truncated before append; flock-based
  mount lock released in tests before reopen.
- rustix fs feature enabled; crate MSRV bumped to 1.87.
- cargo clippy --all-targets --all-features -- -D warnings and
  cargo fmt --check clean; 160 tests passing.

+++ src/tests/rank_roundtrip.rs
@@ -46,7 +46,7 @@ proptest! {
         let mut counts = vec![0u32; m];
         let mut rem = n;
         for c in counts.iter_mut().take(m - 1) {
-            let take = (seed_rank as u64 % (rem + 1)) as u32;
+            let take = (seed_rank % (rem + 1)) as u32;
             *c = take;
             rem -= take as u64;
         }
